Move `tests/test_external_services` -> `tests/test_integrations`
