+++ backend/tests/test_integrations/test_hauki/conftest.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from unittest import mock
 
 import pytest
 
 from tests.factories import OriginHaukiResourceFactory, ReservationUnitFactory
+from tests.helpers import patch_method
 
 if TYPE_CHECKING:
     from tilavarauspalvelu.models import ReservationUnit
@@ -16,9 +16,8 @@ def _force_HaukiAPIClient_to_be_mocked():
     """Force 'HaukiAPIClient.generic' to be mocked in all tests."""
     from tilavarauspalvelu.integrations.opening_hours.hauki_api_client import HaukiAPIClient
 
-    path = HaukiAPIClient.request.__module__ + "." + HaukiAPIClient.request.__qualname__
     exception = NotImplementedError("'HaukiAPIClient.request' must be mocked!")
-    with mock.patch(path, side_effect=exception):
+    with patch_method(HaukiAPIClient.request, side_effect=exception):
         yield
