+++ backend/tests/test_integrations/test_hauki/test_merge_overlapping_time_span_elements.py
@@ -5,7 +5,7 @@ import datetime
 from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
 from tilavarauspalvelu.integrations.opening_hours.time_span_element_utils import merge_overlapping_time_span_elements
 
-from tests.test_external_services.test_hauki.test_reservable_time_spans_client import _get_date
+from tests.test_integrations.test_hauki.test_reservable_time_spans_client import _get_date
 
 # No buffers
