+++ backend/tests/test_integrations/test_hauki/test_override_reservable_with_closed_time_spans.py
@@ -5,7 +5,7 @@ from tilavarauspalvelu.integrations.opening_hours.time_span_element_utils import
     override_reservable_with_closed_time_spans,
 )
 
-from tests.test_external_services.test_hauki.test_reservable_time_spans_client import (
+from tests.test_integrations.test_hauki.test_reservable_time_spans_client import (
     _get_date,
     _get_normalised_time_spans,
     _get_reservable_and_closed_time_spans,
