+++ backend/tests/test_integrations/test_verkkokauppa/test_merchant_types.py
