+++ backend/tests/test_integrations/test_verkkokauppa/test_payment_types.py
