+++ backend/tests/test_integrations/test_verkkokauppa/test_product_types.py
