+++ backend/tests/test_integrations/test_verkkokauppa/test_pruning.py
